[#32] Refactoring test code - Merge all controller test about timeline
